Working on gzip file opening and node age fix.
Adding try/accept logic to try opening gzip file in 'rt' mode,
and then resorting to 'r' mode.

Adding call to ``calc_node_ages`` method for each tree so that all
the node ages are not ``None``!

Removing some print statements.

Multiprocessing is working for Python 2.7.9, but not for 3.4.3
(still a cryptic pickling issue).